requirements.txt
chromadb==1.5.2
huggingface_hub==0.36.2
langchain==0.3.27
langchain-chroma==0.2.6
langchain-community==0.3.31
langchain-core==0.3.83
langchain-huggingface==0.3.1
langchain-ollama==0.3.10
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
ollama==0.6.1
pydantic==2.10.3
pydantic-settings==2.11.0
pydantic_core==2.27.1
pypdf==6.7.5
python-dotenv==1.0.1
sentence-transformers==5.1.2
streamlit==1.50.0
